backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djoser==2.2.2
webcolors==1.13
psycopg2-binary==2.9.9
Pillow==10.3.0
pytest==8.1.1
pytest-django==4.8.0
PyYAML==6.0.1
gunicorn==21.2.0
